requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
fastapi==0.115.11
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
telegramify_markdown==0.5.0